+++ frontend/hooks/calendar/useEventCalendarSync.ts
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { CalendarEvent, Calendar } from '@/utils/calendar/calendar-types';
-import { EventStateActions } from '../types/eventHooks';
+import { EventStateActions } from './types/eventHooks';
 import { isValid } from 'date-fns';
 
 /**
