+++ frontend/hooks/calendar/useEventLoader.ts
@@ -1,10 +1,10 @@
 import { CalendarEvent, Calendar } from '@/utils/calendar/calendar-types';
-import { EventStateActions } from '../types/eventHooks';
+import { EventStateActions } from './types/eventHooks';
 import { useError } from '@/utils/context/ErrorContext';
-import { fetchCalendarEvents } from '../actions';
-import { decryptEventData } from '../../../../utils/calendar/eventEncryption';
-import { validateDecryptedEvent } from '../../../../utils/calendar/eventValidation';
-import { processDecryptedEvent, sortEventsByStartTime } from '../../../../utils/calendar/eventDataProcessing';
+import { fetchCalendarEvents } from '../../app/dashboard/calendar/actions';
+import { decryptEventData } from '../../utils/calendar/eventEncryption';
+import { validateDecryptedEvent } from '../../utils/calendar/eventValidation';
+import { processDecryptedEvent, sortEventsByStartTime } from '../../utils/calendar/eventDataProcessing';
 
 /**
  * Hook for loading and decrypting events
